requirements.txt
flask
google-generativeai
requests
streamlit
SpeechRecognition
PyAudio